Swap definitions of isWhitespace and isSpace
Previously, the behavior of isWhitespace did not match what was stated in the documentation:

Analyse if a char is a white space, that is space, formfeed ('\f'), newline ('\n'), carriage return ('\r'), horizontal tab ('\t'), and vertical tab ('\v')). Returns true if thisChar contains a white space.

Previously, the behavior of isSpace did not match what was stated in the documentation:

Analyse if a char is the space character. Returns true if thisChar contains the space character.

A strict reading of the documentation for isSpace would indicate that it should only match on ASCII value 32 (true space). The change made here will cause it to match only a true space or a tab.