Adding MacAddress, MacAddress8
Adding New Classes MacAddress, MacAddress8 in the style of IPAddress.  Note I didn't include Printable, as I don't understand how that works.  If someone else wants to add it later, they are welcome to do so.  This is my first time contributing to this repository, so please feel free to point out any mistakes.

This library originated on arduino-esp32 (https://github.com/espressif/arduino-esp32/pull/6667).  It was suggested that I contribute to the Arduino core as well :)